모든 계층 구현 완료.. bootstrap적용과 view 데이터 바인딩 해보기

+++ src/main/java/com/doom/service/BoardServiceImpl.java
@@ -7,6 +7,7 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Log4j2
@@ -49,18 +50,39 @@ public class BoardServiceImpl implements BoardService{
 
         log.info("remove......> board_no : {}",board_no);
 
-        int successCnt = boardMapper.deleteBoard(board_no);
-        if (successCnt == 1) {
-            return true;
+        int resultDeleteCnt = getSuccessCntFromDeleteYN(board_no);
+
+        return (resultDeleteCnt==1) ? true : false;
+    }
+
+    private int getSuccessCntFromDeleteYN(Long board_no) {
+        int successCnt = 0;
+        BoardVO boardVO = boardMapper.readBoard(board_no);
+
+        if (boardVO != null && boardVO.getDelete_yn().equals("N")) {
+            successCnt = boardMapper.deleteBoard(board_no);
         }
-        return false;
+        return successCnt;
     }
 
     @Override
     public List<BoardVO> getList() {
 
         log.info("getList.........");
-        List<BoardVO> boardList = boardMapper.getList();
+
+        List<BoardVO> boardList = getBoardList();
+
+        return boardList;
+    }
+
+    private List<BoardVO> getBoardList() {
+        List<BoardVO> boardList = Collections.emptyList(); //예상치못한 NPE를 방지하기 위해 비어있는 list를 선언
+
+        int boardTotalCountNotDelete = boardMapper.readBoardTotalCount(); //delete_yn값이 'N'인 게시글의 갯수
+
+        if (boardTotalCountNotDelete > 0) {
+            boardList = boardMapper.getList(); //삭제되지 않은 개시글 갯수가 존재하면 호출한다.
+        }
         return boardList;
     }
 }
